requirements-dev.txt
pytest>=8.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-httpx>=0.21.0
respx>=0.20.0
mypy>=1.7.0
black>=23.9.0
isort>=5.12.0
pre-commit>=3.5.0
coverage>=7.0.0
ginx>=0.1.6
freezegun>=1.2.0
faker>=19.0.0
responses>=0.23.0
mkdocs>=1.5.0
mkdocs-material>=9.4.0
mkdocstrings[python]>=0.24.0
mkdocs-gen-files>=0.5.0
mkdocs-literate-nav>=0.6.0
mkdocs-section-index>=0.3.0
python-dotenv>=1.0.0
rich>=13.0.0
typer>=0.9.0
